refactor: rename streamlit frontend folder
