--- streamlit/session_bootstrap.py
@@ -1,8 +1,0 @@
-import streamlit as st
-
-from streamlit.cookies_utils import init_cookie_manager_mount
-
-def bootstrap() -> None:
-    """Initialise components that must run before any auth checks."""
-    init_cookie_manager_mount()
-    st.session_state.setdefault("_bootstrapped", True)
