+++ streamlit_app/__init__.py
@@ -1,0 +1,1 @@
+"""OpenSells Streamlit frontend package."""
