+++ streamlit_app/app.py
@@ -3,12 +3,11 @@
 import os
 import streamlit as st
 
-from streamlit.session_bootstrap import bootstrap
+from streamlit_app.auth_utils import ensure_token_and_user, logout_button
+from streamlit_app.cookies_utils import init_cookie_manager_mount
+from streamlit_app.utils import http_client
 
-bootstrap()
-
-from streamlit.auth_utils import ensure_token_and_user, logout_button
-from streamlit.utils import http_client
+init_cookie_manager_mount()
 
 st.set_page_config(page_title="OpenSells — tu motor de prospección y leads", page_icon="🧩")
