+++ streamlit_app/pages/1_Asistente_Virtual.py
@@ -3,14 +3,13 @@ import json
 import streamlit as st
 from dotenv import load_dotenv
 
-from streamlit.session_bootstrap import bootstrap
+from streamlit_app.cache_utils import cached_get, get_openai_client
+from streamlit_app.plan_utils import obtener_plan, tiene_suscripcion_activa, subscription_cta
+from streamlit_app.auth_utils import ensure_token_and_user, logout_button
+from streamlit_app.utils.http_client import get as http_get, post as http_post, health_ok
+from streamlit_app.cookies_utils import init_cookie_manager_mount
 
-bootstrap()
-
-from streamlit.cache_utils import cached_get, get_openai_client
-from streamlit.plan_utils import obtener_plan, tiene_suscripcion_activa, subscription_cta
-from streamlit.auth_utils import ensure_token_and_user, logout_button
-from streamlit.utils.http_client import get as http_get, post as http_post, health_ok
+init_cookie_manager_mount()
 
 st.set_page_config(page_title="Asistente Virtual", page_icon="🤖")
