+++ streamlit_app/pages/3_Mis_Nichos.py
@@ -17,19 +17,18 @@ import hashlib
 from urllib.parse import urlparse
 from dotenv import load_dotenv
 
-from streamlit.session_bootstrap import bootstrap
-
-bootstrap()
-
-from streamlit.cache_utils import (
+from streamlit_app.cache_utils import (
     cached_get,
     cached_post,
     cached_delete,
     limpiar_cache,
 )
-from streamlit.plan_utils import obtener_plan, tiene_suscripcion_activa, subscription_cta
-from streamlit.auth_utils import ensure_token_and_user, logout_button
-from streamlit.utils import http_client
+from streamlit_app.plan_utils import obtener_plan, tiene_suscripcion_activa, subscription_cta
+from streamlit_app.auth_utils import ensure_token_and_user, logout_button
+from streamlit_app.cookies_utils import init_cookie_manager_mount
+from streamlit_app.utils import http_client
+
+init_cookie_manager_mount()
 
 # ── Config ───────────────────────────────────────────
 load_dotenv()
