+++ streamlit_app/pages/7_Suscripcion.py
@@ -5,13 +5,12 @@ import streamlit as st
 import requests
 from dotenv import load_dotenv
 
-from streamlit.session_bootstrap import bootstrap
+from streamlit_app.auth_utils import ensure_token_and_user, logout_button
+from streamlit_app.utils import http_client
+from streamlit_app.plan_utils import obtener_plan, force_redirect
+from streamlit_app.cookies_utils import init_cookie_manager_mount
 
-bootstrap()
-
-from streamlit.auth_utils import ensure_token_and_user, logout_button
-from streamlit.utils import http_client
-from streamlit.plan_utils import obtener_plan, force_redirect
+init_cookie_manager_mount()
 
 load_dotenv()
