+++ streamlit_app/sidebar_utils.py
@@ -1,6 +1,6 @@
 import streamlit as st
 
-from streamlit.cache_utils import limpiar_cache
+from streamlit_app.cache_utils import limpiar_cache
 
 def global_reset_button():
     """Render a sidebar button to clear cache and rerun the app."""
